# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PropFlow API (auth-server + crm-api)",
    "version": "1.0.0",
    "contact": {
      "name": "PropFlow",
      "email": "bionex.product@gmail.com"
    },
    "license": {
      "name": "Proprietary — hiring assignment",
      "url": "https://github.com/your-github/propflow"
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local (npm run dev:auth / docker compose)"
    },
    {
      "url": "https://propflow-auth.up.railway.app",
      "description": "Railway (replace with your deployed URL)"
    },
    {
      "url": "http://localhost:4001",
      "description": "Local (npm run dev:crm / docker compose; via nginx on :8080)"
    },
    {
      "url": "https://propflow-crm.up.railway.app",
      "description": "Railway (replace with your deployed URL)"
    }
  ],
...
  "paths": {
...
    "/platform/tenants": {
      "get": {
        "tags": [
          "Platform"
        ],
        "summary": "Tenants console: every tenant with user counts (SUPER_ADMIN)",
        "description": "Counts only. Property counts come from crm-api `GET /platform/tenant-stats`; the platform console never sees property details, owner phones or chats.",
        "operationId": "listTenants",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "SUPER_ADMIN"
        ],
        "responses": {
          "200": {
            "description": "Tenants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TenantSummary"
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "post": {
        "tags": [
          "Platform"
        ],
        "summary": "Create a tenant and invite its first ADMIN (SUPER_ADMIN)",
        "description": "One transaction: tenant row + ADMIN invite. Returns the invite link for the new admin.",
        "operationId": "createTenant",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "SUPER_ADMIN"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tenant created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTenantResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "Slug or email taken (inline error on the drawer)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "SLUG_TAKEN",
                    "message": "This slug is already taken",
                    "details": {
                      "fields": {
                        "slug": "This slug is already taken"
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
    "/platform/tenants/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Tenant id",
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        }
      ],
      "patch": {
        "tags": [
          "Platform"
        ],
        "summary": "Suspend / reactivate a tenant (SUPER_ADMIN)",
        "description": "Suspending revokes every refresh token of the tenant’s users; logins then fail with 403 ACCOUNT_DISABLED.",
        "operationId": "setTenantActive",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "SUPER_ADMIN"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isActive": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "isActive"
                ]
              },
              "example": {
                "isActive": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tenant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "isActive": {
                      "type": "boolean"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
    "/platform/security-events": {
      "get": {
        "tags": [
          "Platform"
        ],
        "summary": "Security events feed (SUPER_ADMIN)",
        "operationId": "securityEvents",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "SUPER_ADMIN"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "REFRESH_REUSE_DETECTED"
          },
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "beforeId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Keyset pagination cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SecurityEvent"
                      }
                    },
                    "nextBeforeId": {
                      "type": "integer",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
    "/admin/keys": {
      "get": {
        "tags": [
          "Platform"
        ],
        "summary": "Signing keys and their state (SUPER_ADMIN)",
        "operationId": "listKeys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "SUPER_ADMIN"
        ],
        "responses": {
          "200": {
            "description": "Keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SigningKeyInfo"
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
    "/admin/rotate-keys": {
      "post": {
        "tags": [
          "Platform"
        ],
        "summary": "Rotate the signing key pair (SUPER_ADMIN)",
        "description": "Generates a new RSA-2048 pair; new tokens are signed with it immediately. Previous keys are retired but stay in JWKS for `RETIRED_KEY_GRACE_SECONDS`, so access tokens (and invite links) they signed keep verifying — crm-api picks the new `kid` up on first sight without a restart. The new private key is stored AES-256-GCM encrypted with `KEY_ENCRYPTION_KEY` so all auth-server instances can sign with it.",
        "operationId": "rotateKeys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "SUPER_ADMIN"
        ],
        "responses": {
          "201": {
            "description": "Rotated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "kid": {
                      "type": "string"
                    },
                    "retired": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "keys": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SigningKeyInfo"
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT (RS256)",
        "description": "Access token from `/auth/login`, `/auth/refresh`, `/auth/register-tenant` or `/auth/accept-invite`. RS256, header carries `kid`; claims `sub`, `tid`, `role`, `name`, `jti`, `iat`, `exp`, `iss=propflow-auth`, `aud=propflow`. 60 s lifetime in the deployed build. Keep it in memory only."
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "pf_rt",
        "description": "Opaque 256-bit rotating refresh token (7 days). httpOnly; Secure; SameSite=None (web and API are on different sites); Path=/auth. Stored in Redis only as SHA-256. Never sent in a JSON body."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The one error shape returned by both services for every non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "UNAUTHENTICATED",
                  "TOKEN_EXPIRED",
                  "INVALID_TOKEN",
                  "INVALID_CREDENTIALS",
                  "REFRESH_INVALID",
                  "REFRESH_REUSED",
                  "FORBIDDEN_ROLE",
                  "FORBIDDEN_ORIGIN",
                  "NOT_FOUND",
                  "VERSION_CONFLICT",
                  "DUPLICATE_LISTING",
                  "CONFLICT",
                  "EMAIL_TAKEN",
                  "SLUG_TAKEN",
                  "BUSINESS_RULE",
                  "LAST_ADMIN",
                  "IN_USE",
                  "TERMINAL_STATUS",
                  "INVITE_INVALID",
                  "ACCOUNT_DISABLED",
                  "RATE_LIMITED",
                  "INTERNAL",
                  "SERVICE_UNAVAILABLE"
                ],
                "description": "Stable machine-readable code"
              },
              "message": {
                "type": "string",
                "description": "Human-readable message (safe to show in the UI)"
              },
              "details": {
                "description": "Optional, code-specific. For VALIDATION_ERROR it is `ValidationErrorDetails`; for VERSION_CONFLICT it carries the current server copy.",
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ValidationErrorDetails"
                  },
                  {
                    "type": "object",
                    "additionalProperties": true
                  }
                ]
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        }
      },
      "ValidationErrorDetails": {
        "type": "object",
        "properties": {
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Field path → first error message. The web app maps these onto react-hook-form inputs.",
            "example": {
              "priceInr": "Price must be greater than 0",
              "ownerPhone": "Enter a valid 10-digit Indian mobile number"
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        },
        "required": [
          "fields"
        ]
      },
...
      "TenantRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "MANAGER",
          "AGENT"
        ]
      },
...
      "Invite": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "tenantId": {
            "type": "integer"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/TenantRole"
          },
          "invitedBy": {
            "type": "integer",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "acceptedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "state": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REVOKED",
              "EXPIRED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InviteResponse": {
        "type": "object",
        "properties": {
          "invite": {
            "$ref": "#/components/schemas/Invite"
          },
          "inviteUrl": {
            "type": "string",
            "format": "uri",
            "description": "Signed, single-use, 24 h. Also written to the logs (no real email is sent)."
          },
          "inviteToken": {
            "type": "string",
            "description": "RS256 JWT (typ=invite, aud=propflow:invite, jti=invite id)"
          }
        },
        "required": [
          "invite",
          "inviteUrl",
          "inviteToken"
        ],
        "example": {
          "invite": {
            "id": 12,
            "tenantId": 1,
            "email": "rahul@skyline.dev",
            "name": "Rahul Verma",
            "role": "AGENT",
            "invitedBy": 2,
            "expiresAt": "2026-10-02T04:31:10.000Z",
            "acceptedAt": null,
            "revokedAt": null,
            "state": "PENDING",
            "createdAt": "2026-10-01T04:31:10.000Z"
          },
          "inviteUrl": "https://propflow-yourname.web.app/accept-invite?token=eyJhbGciOiJSUzI1NiIs…",
          "inviteToken": "eyJhbGciOiJSUzI1NiIs…"
        }
      },
...
      "TenantSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userCount": {
            "type": "integer"
          },
          "activeUserCount": {
            "type": "integer"
          },
          "usersByRole": {
            "type": "object",
            "properties": {
              "ADMIN": {
                "type": "integer"
              },
              "MANAGER": {
                "type": "integer"
              },
              "AGENT": {
                "type": "integer"
              }
            }
          },
          "pendingInvites": {
            "type": "integer"
          }
        },
        "example": {
          "id": 1,
          "name": "Skyline Realty",
          "slug": "skyline",
          "isActive": true,
          "createdAt": "2026-09-01T10:00:00.000Z",
          "userCount": 4,
          "activeUserCount": 4,
          "usersByRole": {
            "ADMIN": 1,
            "MANAGER": 1,
            "AGENT": 2
          },
          "pendingInvites": 0
        }
      },
      "CreateTenantRequest": {
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "minLength": 3,
            "maxLength": 60
          },
          "adminName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "adminEmail": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "companyName",
          "slug",
          "adminName",
          "adminEmail"
        ],
        "example": {
          "companyName": "Metro Nest",
          "slug": "metro-nest",
          "adminName": "Kiran Rao",
          "adminEmail": "kiran@metronest.in"
        }
      },
      "CreateTenantResponse": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "tenant": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          },
          {
            "$ref": "#/components/schemas/InviteResponse"
          }
        ]
      },
      "SecurityEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string",
            "enum": [
              "LOGIN_SUCCESS",
              "LOGIN_FAILED",
              "LOGIN_LOCKED",
              "REFRESH_REUSE_DETECTED",
              "LOGOUT",
              "LOGOUT_ALL",
              "KEY_ROTATED",
              "TENANT_CREATED",
              "INVITE_CREATED",
              "INVITE_ACCEPTED",
              "USER_UPDATED"
            ]
          },
          "severity": {
            "type": "string",
            "enum": [
              "INFO",
              "WARN",
              "CRITICAL"
            ]
          },
          "userId": {
            "type": "integer",
            "nullable": true
          },
          "tenantId": {
            "type": "integer",
            "nullable": true
          },
          "ip": {
            "type": "string",
            "nullable": true
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "nullable": true
          },
          "meta": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": 88,
          "type": "REFRESH_REUSE_DETECTED",
          "severity": "CRITICAL",
          "userId": 3,
          "tenantId": 1,
          "ip": "49.36.10.4",
          "userAgent": "Mozilla/5.0",
          "requestId": "1d0f…",
          "meta": {
            "familyId": "6b1c…",
            "revokedTokens": 2
          },
          "createdAt": "2026-10-01T05:02:44.000Z"
        }
      },
      "SigningKeyInfo": {
        "type": "object",
        "properties": {
          "kid": {
            "type": "string",
            "description": "RFC 7638 thumbprint (truncated)"
          },
          "alg": {
            "type": "string",
            "enum": [
              "RS256"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "env",
              "rotated"
            ],
            "description": "`env` = bootstrap key from JWT_PRIVATE_KEY; `rotated` = created by /admin/rotate-keys and stored AES-256-GCM encrypted"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "retiredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "signing": {
            "type": "boolean",
            "description": "true for the key currently signing new tokens"
          },
          "publishedUntil": {
            "type": "string",
            "format": "date-time",
            "description": "Retired keys stay in JWKS until their tokens (and invite links) have expired",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "400 — request failed validation against the shared Zod schema",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Request validation failed",
                "details": {
                  "fields": {
                    "email": "Invalid email"
                  },
                  "issues": [
                    {
                      "path": "email",
                      "code": "invalid_string",
                      "message": "Invalid email"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "401 — missing, invalid or expired access token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "missing": {
                "value": {
                  "error": {
                    "code": "UNAUTHENTICATED",
                    "message": "Authentication required"
                  }
                }
              },
              "expired": {
                "summary": "Access token expired → client calls /auth/refresh and retries",
                "value": {
                  "error": {
                    "code": "TOKEN_EXPIRED",
                    "message": "Access token expired"
                  }
                }
              },
              "invalid": {
                "value": {
                  "error": {
                    "code": "INVALID_TOKEN",
                    "message": "Invalid access token"
                  }
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "403 — authenticated, but the role may not do this",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "FORBIDDEN_ROLE",
                "message": "Your role does not allow this action"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "404 — does not exist, belongs to another tenant, or (for an AGENT) is not assigned to you. These cases are deliberately indistinguishable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "NOT_FOUND",
                "message": "Property not found"
              }
            }
          }
        }
      },
...
    },
    "parameters": {
      "XRequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "description": "Optional correlation id (8–64 chars `[A-Za-z0-9._:-]`). Generated when absent and echoed back in the response header; appears on every log line.",
        "schema": {
          "type": "string",
          "example": "web-7f3a9c21"
        }
      }
    },
    "headers": {
      "X-Request-Id": {
        "description": "Correlation id for this request (search logs with it)",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    }
  },
...
}